Add init callback to gen_ga behavior
After creating a new GA instance, the Module:init/1 is called to
generate the initial population. It takes the population size as its
only parameter. Simplifies initialization and future proofs the gen_ga
bahavior so that the GA can be reinitialized easily if that is
something we end up needing. This keeps code for generating random
genotypes in the same place as all the other genotype-specific
code (mutation, crossover, and evaluation).